Remplacer float par double et utiliser size_t dans lab01b

+++ laboratoires/lab01-pointeurs/solution/lab01b/fonctions.cpp
@@ -2,21 +2,21 @@
 
 using namespace std;
 
-void demanderNombres(float nombres[], int indexMin, int indexMax) {
-    for (int i = indexMin; i <= indexMax; i++) {
+void demanderNombres(double nombres[], size_t indexMin, size_t indexMax) {
+    for (size_t i = indexMin; i <= indexMax; i++) {
         cout << "Entrez le nombre #" << (i + 1) << ": ";
         cin >> nombres[i];
     }
 }
 
-float calculerMoyenne(float nombres[], int nombreValeurs) {
-    float somme = 0;
-    for (int i = 0; i < nombreValeurs; i++) {
+double calculerMoyenne(double nombres[], size_t nombreValeurs) {
+    double somme = 0;
+    for (size_t i = 0; i < nombreValeurs; i++) {
         somme += nombres[i];
     }
     return somme / nombreValeurs;
 }
 
-void afficherMoyenne(int nombreValeurs, float moyenne) {
+void afficherMoyenne(size_t nombreValeurs, double moyenne) {
     cout << "La moyenne des " << nombreValeurs << " nombres est " << moyenne << "." << endl;
 }
